New folder/AppGenesisCompilerAPI/requirements.txt
djangorestframework
environs
django-allauth
djangorestframework-simplejwt
dj-rest-auth
google-auth
google-auth-oauthlib 
google-auth-httplib2